Push 6
Updated bjork jump cycle to be smoother. Testing on open processing and was expereincing serious lag, so I optimized the amount of stars being drawn in the background and called all load functions inside preload to mitigate.

Added enemy un bombs that push bjork off canvas to increase difficulty of the game.